update profiler, verify code is working

+++ src/utils/math/blas/Constants.h
@@ -11,7 +11,7 @@
 
 namespace MyBLAS {
 
-using NumericType = long double;
+using NumericType = float;
 
 }
